move application-crypto tests into seperate crate
