+++ primitives/application-crypto/Cargo.toml
@@ -12,10 +12,6 @@ serde = { version = "1.0.101", optional = true, features = ["derive"] }
 sp-std = { path = "../std", default-features = false }
 runtime-io = { package = "sp-io", path = "../../primitives/sr-io", default-features = false }
 
-[dev-dependencies]
-test-client = { package = "substrate-test-runtime-client", path = "../../test/utils/runtime/client" }
-sp-runtime = { path = "../../primitives/runtime" }
-
 [features]
 default = [ "std" ]
 std = [ "full_crypto", "primitives/std", "codec/std", "serde", "sp-std/std", "runtime-io/std" ]
