+++ primitives/application-crypto/test/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "sp-application-crypto-test"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+description = "Integration tests for application-crypto"
+publish = false
+
+[dependencies]
+primitives = { package = "sp-core", path = "../../core", default-features = false }
+# codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false, features = ["derive"] }
+# serde = { version = "1.0.101", optional = true, features = ["derive"] }
+# sp-std = { path = "../std", default-features = false }
+# runtime-io = { package = "sp-io", path = "../../primitives/sr-io", default-features = false }
+test-client = { package = "substrate-test-runtime-client", path = "../../../test/utils/runtime/client" }
+sp-runtime = { path = "../../runtime" }
